refactor: add workflow shell presenter API
