+++ ui/dockwidget/__init__.py
@@ -1,6 +1,15 @@
-"""Dock widget components for the qfit wizard shell."""
+"""Dock widget components for qfit workflow dock shells."""
 
 from .stepper_bar import STEPPER_LABELS, STEPPER_STATES, StepperBar
-from .wizard_shell_presenter import WizardShellPresenter
+from .workflow_shell_presenter import WorkflowShellPresenter
 
-__all__ = ["STEPPER_LABELS", "STEPPER_STATES", "StepperBar", "WizardShellPresenter"]
+WizardShellPresenter = WorkflowShellPresenter
+"""Compatibility alias for pre-#805 wizard shell presenter imports."""
+
+__all__ = [
+    "STEPPER_LABELS",
+    "STEPPER_STATES",
+    "StepperBar",
+    "WorkflowShellPresenter",
+    "WizardShellPresenter",
+]
